Первое задание - исправление на double

+++ src/Main.java
@@ -24,7 +24,7 @@ public class Main {
 
     public static void case3(){
         int number= 10500;
-        int result = number/10/10;
+        double result = number/10d/10d;
         System.out.println("3. \n (10500/10)/10 = "+result);
     }
